Initial approach with starter library to help out
