+++ build.gradle
@@ -22,8 +22,9 @@ dependencies {
 	implementation 'com.graphql-java-kickstart:graphql-java-tools:6.0.2'
 	implementation 'com.graphql-java-kickstart:graphql-spring-boot-starter:7.0.0'
 	runtimeOnly 'com.graphql-java-kickstart:graphiql-spring-boot-starter:7.0.0'
-	implementation 'org.springframework.boot:spring-boot-starter-web'
 	implementation 'org.springframework.cloud:spring-cloud-gcp-starter'
+	implementation 'org.springframework.boot:spring-boot-starter-web'
+	implementation 'org.springframework.cloud:spring-cloud-gcp-starter-data-firestore'
 	testImplementation('org.springframework.boot:spring-boot-starter-test') {
 		exclude group: 'org.junit.vintage', module: 'junit-vintage-engine'
 	}
