+++ configmap.yaml
@@ -1,0 +1,18 @@
+apiVersion: v1
+kind: ConfigMap
+metadata:
+  name: firebase-service-account-cm
+data:
+  firebase-service-account.json: |
+    {
+    "type": "service_account",
+    "project_id": "$FIREBASE_PROJECT_ID",
+    "private_key_id": "$FIREBASE_SERVICE_ACCOUNT_PRIVATE_KEY_ID",
+    "private_key": "$FIREBASE_SERVICE_ACCOUNT_PRIVATE_KEY",
+    "client_email": "FIREBASE_SERVICE_ACCOUNT_CLIENT_EMAIL",
+    "client_id": "",
+    "auth_uri": "$FIREBASE_SERVICE_ACCOUNT_AUTH_URI",
+    "token_uri": "$FIREBASE_SERVICE_ACCOUNT_TOKEN_URI",
+    "auth_provider_x509_cert_url": "$FIREBASE_SERVICE_ACCOUNT_AUTH_PROVIDER_X509_CERT_URL",
+    "client_x509_cert_url": "$FIREBASE_SERVICE_ACCOUNT_CLIENT_X509_CERT_URL"
+    }
