+++ kustomization.yaml
@@ -3,3 +3,4 @@ kind: Kustomization
 resources:
 - deployment.yaml
 - service.yaml
+- configmap.yaml
