+++ src/main/java/com/thirdwavelist/barista/entity/Cafe.java
@@ -1,0 +1,27 @@
+package com.thirdwavelist.barista.entity;
+
+public class Cafe {
+    private String id;
+    private String name;
+
+    public Cafe(String id, String name) {
+        this.id = id;
+        this.name = name;
+    }
+
+    public String getId() {
+        return id;
+    }
+
+    public void setId(String id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+}
