--- src/main/java/com/thirdwavelist/barista/model/BrewMethod.java
@@ -1,10 +1,0 @@
-package com.thirdwavelist.barista.model;
-
-public enum BrewMethod {
-    ESPRESSO,
-    AEROPRESS,
-    POUROVER,
-    COLDBREW,
-    SYPHON,
-    FULLIMMERSION,
-}
