--- src/main/java/com/thirdwavelist/barista/model/Currency.java
@@ -1,8 +1,0 @@
-package com.thirdwavelist.barista.model;
-
-public enum Currency {
-    HUF,
-    USD,
-    EUR,
-    GBP,
-}
