--- src/main/java/com/thirdwavelist/barista/model/Machine.java
@@ -1,13 +1,0 @@
-package com.thirdwavelist.barista.model;
-
-public class Machine {
-    private String name;
-
-    public Machine(String name) {
-        this.name = name;
-    }
-
-    public String getName() {
-        return name;
-    }
-}
