--- src/main/java/com/thirdwavelist/barista/model/OriginType.java
@@ -1,6 +1,0 @@
-package com.thirdwavelist.barista.model;
-
-public enum OriginType {
-    SINGLE,
-    BLEND,
-}
