--- src/main/java/com/thirdwavelist/barista/model/RoastLevel.java
@@ -1,7 +1,0 @@
-package com.thirdwavelist.barista.model;
-
-public enum RoastLevel {
-    LIGHT,
-    MEDIUM,
-    DARK,
-}
