+++ src/main/java/com/thirdwavelist/barista/query/CafeQueryResolver.kt
@@ -1,6 +1,6 @@
 package com.thirdwavelist.barista.query
 
-import com.thirdwavelist.barista.model.Cafe
+import com.thirdwavelist.barista.entity.Cafe
 
 import com.thirdwavelist.barista.service.CafeService
 
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Service
 
 @Service
 class CafeQueryResolver(private val cafeService: CafeService) : GraphQLQueryResolver {
-    fun cafe(): Collection<Cafe> {
-        return cafeService.cafe
+    fun cafe(id: String): Cafe {
+        return cafeService.getCafe(id)
     }
 }
