+++ src/main/java/com/thirdwavelist/barista/repository/CafeRepository.java
@@ -1,0 +1,7 @@
+package com.thirdwavelist.barista.repository;
+
+import com.thirdwavelist.barista.entity.CafeDto;
+import org.springframework.cloud.gcp.data.firestore.FirestoreReactiveRepository;
+
+public interface CafeRepository extends FirestoreReactiveRepository<CafeDto> {
+}
