+++ src/main/resources/application.properties
@@ -1,1 +1,4 @@
+# Use this setting if you want to manually specify a GCP Project instead of inferring
+# from your machine's environment.
+# spring.cloud.gcp.firestore.project-id={YOUR_PROJECT_ID}
